improvements to speed command + gasPrice as double in transaction template

+++ src/main/java/be/e_contract/ethereum/tool/TransactionTemplate.java
@@ -31,7 +31,7 @@ public class TransactionTemplate {
     public double value;
 
     // unit: gwei
-    public int gasPrice;
+    public double gasPrice;
 
     public long nonce;
